Move mock classes into `export/testing`.
