+++ src/compy/index.js
@@ -4,8 +4,6 @@
 export * from '#x/BaseComponent';
 export * from '#x/BaseRootComponent';
 export * from '#x/ControlContext';
-export * from '#x/MockComponent';
-export * from '#x/MockRootComponent';
 export * from '#x/Names';
 export * from '#x/RootControlContext';
 export * from '#x/TemplAggregateComponent';
