+++ src/compy/package.json
@@ -6,7 +6,8 @@
   "license": "Apache-2.0",
 
   "exports": {
-    ".": "./index.js"
+    ".": "./index.js",
+    "./testing": "./export/testing/index.js"
   },
   "imports": {
     "#x/*": "./export/*.js",
