+++ src/webapp-builtins/tests/PathRouter.test.js
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { PathKey } from '@this/collections';
-import { MockComponent, MockRootComponent } from '@this/compy';
+import { MockComponent, MockRootComponent } from '@this/compy/testing';
 import { DispatchInfo } from '@this/net-util';
 import { PathRouter } from '@this/webapp-builtins';
